+++ backend/swagger/swagger.js
@@ -10,9 +10,9 @@ const doc = {
    * push할 때는 host: 3001로 해주삼 !
    * (주의) main 배포는...3000으로 해야하나... 여기까지 알아보기는 마음이 지금 아픔
    */
-  //host: 'localhost:3000',
-  host: '43.202.208.226:3001',
-    schemes: ['http'],
+  host: "localhost:3001",
+  //host: '43.202.208.226:3001',
+  schemes: ["http"],
   securityDefinitions: {
     Bearer: {
       type: "apiKey",
