+++ backend/swagger/swagger_output.json
@@ -5,11 +5,9 @@
     "description": "오늘도 화이팅",
     "version": "1.0.0"
   },
-  "host": "localhost:3000",
+  "host": "localhost:3001",
   "basePath": "/",
-  "schemes": [
-    "http"
-  ],
+  "schemes": ["http"],
   "securityDefinitions": {
     "Bearer": {
       "type": "apiKey",
@@ -21,9 +19,7 @@
   "paths": {
     "/tarot/guide": {
       "get": {
-        "tags": [
-          "Tarot"
-        ],
+        "tags": ["Tarot"],
         "summary": "가이드라인 불러오기",
         "description": "운 종류와 뽑는 사람 수를 전달하면 가이드라인(content)과 타로 마스터 이름(master_name)을 반환함",
         "parameters": [
@@ -117,9 +113,7 @@
     },
     "/tarot/card/info": {
       "post": {
-        "tags": [
-          "Tarot"
-        ],
+        "tags": ["Tarot"],
         "summary": "카드 정보 조회",
         "description": "카드 정보를 정수형으로 전달하면 해당 카드의 정보를 반환함",
         "parameters": [
@@ -211,9 +205,7 @@
     },
     "/poll/create": {
       "get": {
-        "tags": [
-          "Tarot"
-        ],
+        "tags": ["Tarot"],
         "summary": "타로 시작 시 Poll(임시저장) → 타로 시작 할 경우 뽑은 카드와 결과 저장을 구별할 Poll Table",
         "description": "뽑은 카드 결과 저장 및 총 결과 저장 시 사용됨",
         "responses": {
@@ -303,9 +295,7 @@
     },
     "/result/save": {
       "post": {
-        "tags": [
-          "Result"
-        ],
+        "tags": ["Result"],
         "summary": "종합 결과 저장",
         "description": "사용자의 질문과 관련된 타로 종합 결과 및 뽑은 카드 정보를 저장합니다. 사용자의 운(luck)에 따라 카드의 수가 결정됩니다.",
         "parameters": [
@@ -381,11 +371,7 @@
                 },
                 "cardIds": {
                   "type": "array",
-                  "example": [
-                    1,
-                    2,
-                    3
-                  ],
+                  "example": [1, 2, 3],
                   "items": {
                     "type": "number"
                   }
@@ -442,9 +428,7 @@
     },
     "/user/signup": {
       "post": {
-        "tags": [
-          "User"
-        ],
+        "tags": ["User"],
         "summary": "회원가입",
         "description": "회원가입 하는 유저의 아이디 중복검사 및 가입 하는 유저의 정보를 데이터베이스에 저장",
         "parameters": [
@@ -545,9 +529,7 @@
     },
     "/user/login": {
       "post": {
-        "tags": [
-          "User"
-        ],
+        "tags": ["User"],
         "summary": "로그인",
         "description": "아이디와 비밀번호를 이용하여 로그인을 수행하고, 성공 시 액세스 토큰과 리프레시 토큰을 발급",
         "parameters": [
@@ -640,9 +622,7 @@
     },
     "/user/logout": {
       "post": {
-        "tags": [
-          "User"
-        ],
+        "tags": ["User"],
         "summary": "로그아웃",
         "description": "현재 로그인된 사용자의 로그아웃을 수행",
         "responses": {
@@ -679,9 +659,7 @@
     },
     "/mypage/detail": {
       "get": {
-        "tags": [
-          "MyPage"
-        ],
+        "tags": ["MyPage"],
         "summary": "결과 리스트에서 선택한 결과(poll_id) 상세 조회",
         "description": "결과 리스트 중 선택한 결과(Poll_id)를 통해 해당 결과를 상세조회한다.",
         "parameters": [
@@ -777,9 +755,7 @@
     },
     "/mypage/delete": {
       "delete": {
-        "tags": [
-          "MyPage"
-        ],
+        "tags": ["MyPage"],
         "summary": "결과 리스트에서 선택한 결과 및 카드 삭제",
         "description": "결과 리스트 중 선택한 결과(Poll_id)가 삭제된다. -> CASCADE로 result, card 테이블의 관련 데이터도 삭제",
         "parameters": [
@@ -879,9 +855,7 @@
     },
     "/test/testgptsplit": {
       "get": {
-        "tags": [
-          "Test"
-        ],
+        "tags": ["Test"],
         "summary": "GPT-3 테스트",
         "description": "GPT-3를 테스트 해봅시다 :>",
         "responses": {
@@ -902,9 +876,7 @@
     },
     "/secret/MySQL": {
       "get": {
-        "tags": [
-          "Test"
-        ],
+        "tags": ["Test"],
         "summary": "시크릿 매니저를 통한 MySQL 정보 조회",
         "description": "SecretsManager Router Test -> MySQL",
         "responses": {
@@ -928,9 +900,7 @@
     },
     "/token/refresh": {
       "post": {
-        "tags": [
-          "Token"
-        ],
+        "tags": ["Token"],
         "summary": "리프레시 토큰을 사용하여 새로운 엑세스 토큰과 리프레시 토큰 발급",
         "description": "유효한 리프레시 토큰을 제공받아 새로운 엑세스 토큰과 리프레시 토큰을 발급합니다.",
         "parameters": [
@@ -992,9 +962,7 @@
     },
     "/stream/": {
       "post": {
-        "tags": [
-          "Stream"
-        ],
+        "tags": ["Stream"],
         "summary": "타로 결과 GPT 요청",
         "description": "타로 결과를 API에 요청하고 결과를 반환함",
         "parameters": [
@@ -1010,11 +978,7 @@
               "properties": {
                 "cards": {
                   "type": "array",
-                  "example": [
-                    1,
-                    2,
-                    3
-                  ],
+                  "example": [1, 2, 3],
                   "items": {
                     "type": "number"
                   }
